Refactor extension.ts: Modularize command registration and graph handling
- Removed unused imports and code related to cycle analysis and command registration.
- Introduced GraphController and CommandRegistry for better separation of concerns.
- Simplified the activation process by delegating command registration to CommandRegistry.
- Enhanced error handling and logging during command registration and graph generation.
- Improved structure for handling graph preview and cycle analysis.
